Rename internal/api/aignite package to ignite
The aignite name was a workaround for an unanchored ignite/ gitignore rule
that swallowed any directory named ignite at any depth. This repo's rules
are anchored to the Ignite runtime work directories (/ignite/ and
*/ignite/), so a source package named ignite commits cleanly - verified
with git check-ignore on the new paths.
