+++ kinotic-core/src/main/java/org/kinotic/core/internal/api/event/DefaultEventBusService.java
@@ -16,7 +16,7 @@ import org.kinotic.core.api.event.EventBusService;
 import org.kinotic.core.api.event.EventConstants;
 import org.kinotic.core.api.event.EventConsumer;
 import org.kinotic.core.api.event.ListenerStatus;
-import org.kinotic.core.internal.api.aignite.KinoticIgniteClusterManager;
+import org.kinotic.core.internal.api.ignite.KinoticIgniteClusterManager;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import reactor.core.publisher.Flux;
