+++ kinotic-core/src/main/java/org/kinotic/core/internal/api/ignite/Observer.java
@@ -1,6 +1,6 @@
 
 
-package org.kinotic.core.internal.api.aignite;
+package org.kinotic.core.internal.api.ignite;
 
 import io.vertx.core.Handler;
