+++ kinotic-core/src/main/java/org/kinotic/core/internal/config/KinoticVertxConfig.java
@@ -11,8 +11,8 @@ import io.vertx.core.spi.cluster.ClusterManager;
 import org.apache.ignite.Ignite;
 import org.kinotic.core.api.config.KinoticProperties;
 import org.kinotic.core.api.event.Event;
-import org.kinotic.core.internal.api.aignite.KinoticIgniteClusterManager;
 import org.kinotic.core.internal.api.event.EventMessageCodec;
+import org.kinotic.core.internal.api.ignite.KinoticIgniteClusterManager;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.annotation.Bean;
